IGNITE-16566 Move ErrorGroups to api module
