+++ modules/api/src/main/java/org/apache/ignite/lang/ErrorGroup.java
