+++ modules/api/src/main/java/org/apache/ignite/lang/ErrorGroups.java
